+++ console/src_b/game_b.c
@@ -43,13 +43,13 @@ void FreeGame(Game *g) {
     }
 }
 
-int **MakeBoard() {
-    int **board = NULL;
-    board = malloc(4 * sizeof(int *));
+unsigned short int **MakeBoard() {
+    unsigned short int **board = NULL;
+    board = malloc(4 * sizeof(unsigned short int *));
     if (board == NULL) { MALLOC_FAIL }
     for (int i = 0; i < 4; i++) {
         board[i] = NULL;
-        board[i] = malloc(4 * (sizeof(int)));
+        board[i] = malloc(4 * (sizeof(unsigned short int)));
         if (board[i] == NULL) {
             FreeBoard(board);
             MALLOC_FAIL
@@ -59,7 +59,7 @@ int **MakeBoard() {
     return board;
 }
 
-void InitBoard(int **board) {
+void InitBoard(unsigned short int **board) {
     if (board == NULL) {
         printf("could not find board to initialize\n");
         FAIL_OUT
@@ -71,7 +71,7 @@ void InitBoard(int **board) {
     }
 }
 
-void FreeBoard(int **board) {
+void FreeBoard(unsigned short int **board) {
     if (board != NULL) {
         for (int i = 0; i < 4; i++) {
             if (board[i] != NULL) { free(board[i]); }
@@ -80,12 +80,12 @@ void FreeBoard(int **board) {
     }
 }
 
-int **CopyBoard(int **board) {
+unsigned short int **CopyBoard(unsigned short int **board) {
     if (board == NULL) {
         printf("could not find board to copy\n");
         FAIL_OUT
     }
-    int **res = MakeBoard();
+    unsigned short int **res = MakeBoard();
     for (int i = 0; i < 4; i++) {
         for (int j = 0; j < 4; j++) {
             res[i][j] = board[i][j];
@@ -138,12 +138,12 @@ void SaveGame(Game *g) {
     if (isFOpen(fp, fname)) {
         for (int i = 0; i < 4; i++) {
             for (int j = 0; j < 4; j++) {
-                fprintf(fp, "%d ", g->board[i][j]);
+                fprintf(fp, "%hu ", g->board[i][j]);
             }
             fprintf(fp, "\n");
         }
-        fprintf(fp, "%d ", g->score);
-        fprintf(fp, "%d ", g->free_tiles);
+        fprintf(fp, "%hu ", g->score);
+        fprintf(fp, "%hu ", g->free_tiles);
         fclose(fp);
         printf("partie sauvegardee\n");
     } else {
@@ -162,18 +162,18 @@ int LoadGame(Game *g) {
     char *fname = "console/data_b/save.txt";
     fp = fopen(fname, "r");
     if (isFOpen(fp, fname)) {
-        int tmp;
+        unsigned short int tmp;
         InitGame(g);
         for (int i = 0; i < 4; i++) {
             for (int j = 0; j < 4; j++) {
-                fscanf(fp, "%d ", &tmp);
+                fscanf(fp, "%hu ", &tmp);
                 g->board[i][j] = tmp;
             }
             fscanf(fp, "\n");
         }
-        fscanf(fp, "%d ", &tmp);
+        fscanf(fp, "%hu ", &tmp);
         g->score = tmp;
-        fscanf(fp, "%d ", &tmp);
+        fscanf(fp, "%hu ", &tmp);
         g->free_tiles = tmp;
         printf("partie chargee\n");
         fclose(fp);
@@ -183,7 +183,7 @@ int LoadGame(Game *g) {
     }
 }
 
-void SpawnTiles(Game *g, int val, int n) {
+void SpawnTiles(Game *g, unsigned short int val, int n) {
     if (g == NULL || (val != 2 && val != 4) || (n != 1 && n != 2)) {
         printf("could not find game to spawn tiles, or other invalid param\n");
         FAIL_OUT
@@ -320,13 +320,13 @@ void Fuse(Game *g, int *hasFused) {
     }
 }
 
-void Rotate(int **board, int n) {
+void Rotate(unsigned short int **board, int n) {
     if (board == NULL || n < 1 || n > 3) {
         printf("could not find board to rotate, or other invalid param\n");
         FAIL_OUT
     }
     for (int a = 0; a < n; a++) {
-        int **aux = CopyBoard(board); // original copy of board
+        unsigned short int **aux = CopyBoard(board); // original copy of board
         for (int i = 0; i < 4; i++) {
             for (int j = 0; j < 4; j++) {
                 board[j][3 - i] = aux[i][j];
